add tests for accounts and users views

+++ password_manager/users/views.py
@@ -146,6 +146,9 @@ class LogoutView(APIView):
         },
     )
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials were not provided."},
+                            status=status.HTTP_401_UNAUTHORIZED)
         if hasattr(request.user, 'auth_token'):
             request.user.auth_token.delete()
             return Response({"message": "Successfully logged out."}, status=status.HTTP_200_OK)
